requirements.txt
openai
pandas
sqlalchemy
requests
python-dotenv
